Alterei o tipo do id do professor para Long

+++ src/main/java/com/md/escola/models/Professor.java
@@ -6,8 +6,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
 
 @NoArgsConstructor
 @AllArgsConstructor
@@ -17,7 +15,7 @@ public class Professor  {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     String data_admissao;
     @OneToOne
     @JoinColumn(name = "pessoa_id")
